# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "面向生产环境的本地文件扫描、持久化上传队列与 rclone 调度 API。",
        "title": "Rclone Sync Hub API",
        "contact": {
            "name": "Project repository",
            "url": "https://github.com/zane626/rclone-sync-hub-go-backend"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
        "/api/analytics/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回概览数字、按状态/按监听文件夹/按时间趋势的图表数据，以及最近任务与失败任务列表，供数据分析页一次拉取",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "数据分析仪表盘",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "趋势图天数，默认 7",
                        "name": "days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rclone-sync-hub_internal_service.DashboardData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/audit-logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operations"
                ],
                "summary": "获取操作审计日志",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "用户名和密码",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "获取当前用户",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/fs/subdirs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据传入的本地路径，返回该路径下所有一级子文件夹（不包含文件）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "filesystem"
                ],
                "summary": "获取指定路径下的子文件夹",
                "parameters": [
                    {
                        "type": "string",
                        "description": "本地起始路径，如 /volumes",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.FSSubDirListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/rclone/configs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "调用 `rclone listremotes --long` 获取 remote 名称与类型，不读取配置密钥",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rclone"
                ],
                "summary": "获取 rclone remote 配置列表",
                "responses": {
                    "200": {
                        "description": "items 为 remote 列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error 为错误信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/scan-runs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operations"
                ],
                "summary": "获取扫描历史",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "监听目录 ID；0 为全部",
                        "name": "watch_folder_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_api.FSSubDir": {
            "type": "object",
            "properties": {
                "has_sub_dirs": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "internal_api.FSSubDirListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.FSSubDir"
                    }
                }
            }
        },
        "internal_api.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 1024
                },
                "username": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
...
        "rclone-sync-hub_internal_model.FileRecord": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "fileHash": {
                    "description": "可选，内容哈希用于强校验",
                    "type": "string"
                },
                "fileModTime": {
                    "type": "string"
                },
                "fileSize": {
                    "type": "integer"
                },
                "fingerprint": {
                    "description": "size + mtime 的快速元数据指纹",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastSeenAt": {
                    "type": "string"
                },
                "localPath": {
                    "type": "string"
                },
                "missingAt": {
                    "description": "最近一次完整扫描确认文件已不存在",
                    "type": "string"
                },
                "relativePath": {
                    "description": "相对扫描根目录的路径",
                    "type": "string"
                },
                "remoteName": {
                    "type": "string"
                },
                "remotePath": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uploadedAt": {
                    "type": "string"
                },
                "watchFolderID": {
                    "type": "integer"
                }
            }
        },
...
        "rclone-sync-hub_internal_model.UploadTask": {
            "type": "object",
            "properties": {
                "accumulatedFailures": {
                    "description": "累计失败次数（便于分析）",
                    "type": "integer"
                },
                "cancelRequestedAt": {
                    "description": "运行中任务的取消请求",
                    "type": "string"
                },
                "canceledAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "durationSeconds": {
                    "description": "时间维度",
                    "type": "integer"
                },
                "errorMsg": {
                    "description": "最后一次错误信息",
                    "type": "string"
                },
                "fileFingerprint": {
                    "description": "创建任务时的文件元数据指纹",
                    "type": "string"
                },
                "fileName": {
                    "description": "文件名",
                    "type": "string"
                },
                "fileRecord": {
                    "description": "关联（不参与表结构，仅查询用）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/rclone-sync-hub_internal_model.FileRecord"
                        }
                    ]
                },
                "fileRecordID": {
                    "type": "integer"
                },
                "fileSize": {
                    "description": "文件信息",
                    "type": "integer"
                },
                "finishedAt": {
                    "description": "上传结束时间",
                    "type": "string"
                },
                "heartbeatAt": {
                    "description": "最近一次租约心跳",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "idempotencyKey": {
                    "description": "nullable，兼容旧数据并防止重复建任务",
                    "type": "string"
                },
                "lastProgressAt": {
                    "description": "最近一次进度上报时间",
                    "type": "string"
                },
                "lastStatusAt": {
                    "description": "最近一次状态变更时间",
                    "type": "string"
                },
                "leaseExpiresAt": {
                    "description": "租约到期时间",
                    "type": "string"
                },
                "leaseOwner": {
                    "description": "当前 worker 实例",
                    "type": "string"
                },
                "localPath": {
                    "description": "文件本地路径",
                    "type": "string"
                },
                "log": {
                    "description": "日志与分析",
                    "type": "string"
                },
                "nextRetryAt": {
                    "description": "下次允许领取时间",
                    "type": "string"
                },
                "priority": {
                    "description": "数值越大优先级越高",
                    "type": "integer"
                },
                "progress": {
                    "description": "上传进度（0-100）",
                    "type": "number"
                },
                "remoteName": {
                    "description": "上传网盘（remote 名称）",
                    "type": "string"
                },
                "remotePath": {
                    "description": "上传路径",
                    "type": "string"
                },
                "retryCount": {
                    "description": "已领取执行次数（包含首次）",
                    "type": "integer"
                },
                "speed": {
                    "description": "最近一次上报的速度（bytes/s）",
                    "type": "integer"
                },
                "startedAt": {
                    "description": "上传开始时间",
                    "type": "string"
                },
                "status": {
                    "description": "状态与进度",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "watchFolderID": {
                    "description": "归属与标识",
                    "type": "integer"
                },
                "watchFolderName": {
                    "description": "所属监听文件夹名称快照",
                    "type": "string"
                }
            }
        },
...
        "rclone-sync-hub_internal_service.DashboardData": {
            "type": "object",
            "properties": {
                "by_status": {
                    "description": "ByStatus 按状态分布（饼图/柱状图）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rclone-sync-hub_internal_service.StatusItemVO"
                    }
                },
                "by_time": {
                    "description": "ByTime 按日趋势（折线图）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rclone-sync-hub_internal_service.TimeItemVO"
                    }
                },
                "by_watch_folder": {
                    "description": "ByWatchFolder 按监听文件夹分布（表格+柱状图）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rclone-sync-hub_internal_service.WatchFolderItemVO"
                    }
                },
                "items": {
                    "description": "Items 列表类数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/rclone-sync-hub_internal_service.DashboardItemsVO"
                        }
                    ]
                },
                "overview": {
                    "description": "Overview 概览数字（卡片）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/rclone-sync-hub_internal_service.OverviewVO"
                        }
                    ]
                }
            }
        },
        "rclone-sync-hub_internal_service.DashboardItemsVO": {
            "type": "object",
            "properties": {
                "failed_tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rclone-sync-hub_internal_model.UploadTask"
                    }
                },
                "recent_tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rclone-sync-hub_internal_model.UploadTask"
                    }
                }
            }
        },
        "rclone-sync-hub_internal_service.OverviewVO": {
            "type": "object",
            "properties": {
                "recent_24h_completed": {
                    "type": "integer"
                },
                "recent_24h_failed": {
                    "type": "integer"
                },
                "task_canceled": {
                    "type": "integer"
                },
                "task_failed": {
                    "type": "integer"
                },
                "task_paused": {
                    "type": "integer"
                },
                "task_pending": {
                    "type": "integer"
                },
                "task_running": {
                    "type": "integer"
                },
                "task_success": {
                    "type": "integer"
                },
                "task_total": {
                    "type": "integer"
                },
                "uploaded_bytes_total": {
                    "type": "integer"
                },
                "uploaded_files_total": {
                    "type": "integer"
                },
                "watch_folder_count": {
                    "type": "integer"
                }
            }
        },
        "rclone-sync-hub_internal_service.StatusItemVO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "rclone-sync-hub_internal_service.TimeItemVO": {
            "type": "object",
            "properties": {
                "completed_count": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "failed_count": {
                    "type": "integer"
                },
                "uploaded_bytes": {
                    "type": "integer"
                }
            }
        },
        "rclone-sync-hub_internal_service.WatchFolderItemVO": {
            "type": "object",
            "properties": {
                "canceled_count": {
                    "type": "integer"
                },
                "failed_count": {
                    "type": "integer"
                },
                "paused_count": {
                    "type": "integer"
                },
                "pending_count": {
                    "type": "integer"
                },
                "running_count": {
                    "type": "integer"
                },
                "success_count": {
                    "type": "integer"
                },
                "task_count": {
                    "type": "integer"
                },
                "uploaded_bytes": {
                    "type": "integer"
                },
                "uploaded_files": {
                    "type": "integer"
                },
                "watch_folder_id": {
                    "type": "integer"
                },
                "watch_folder_name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}